src/backend/pyproject.toml
[project]
name = "aura"
version = "1.0.0"
description = "Multi-tenant AI-powered resume screening tool for HR teams"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn==0.30.0",
    "pydantic==2.9.0",
    "pydantic-settings==2.5.0",
    "sqlalchemy==2.0.35",
    "asyncpg==0.29.0",
    "redis==5.1.0",
    "qdrant-client==1.11.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-text-splitters>=0.3.0",
    "pymupdf>=1.24.0",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
